+++ AoE2ScenarioRms/util/xs_util.py
@@ -11,3 +11,8 @@ class XsUtil:
     def read(path: str) -> str:
         with (Path(__file__).parent.parent / 'xs' / path).open() as file:
             return file.read()
+
+    @staticmethod
+    def format_name(name: str):
+        xs_name = ' '.join(filter(lambda e: e, name.split(' ')))
+        return xs_name.upper().replace(' ', '_')